test: keep the sudo attribution and refusal tests on Unix
The sudo marker counts on Unix only, so the two tests that set it and
expect the sudo path have no Windows case. They move to Unix-only files.
