+++ internal/audit/cloud_sink_test.go
@@ -3,7 +3,6 @@ package audit
 import (
 	"context"
 	"database/sql"
-	"os/user"
 	"sync"
 	"testing"
 	"time"
@@ -11,7 +10,6 @@ import (
 	controltowerv1 "buf.build/gen/go/safedep/api/protocolbuffers/go/safedep/messages/controltower/v1"
 	servicev1 "buf.build/gen/go/safedep/api/protocolbuffers/go/safedep/services/controltower/v1"
 	"github.com/safedep/dry/cloud/endpointsync"
-	"github.com/safedep/pmg/internal/platform"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -412,50 +410,6 @@ func TestCloudSinkSetsKubernetesContextOnSessionSummary(t *testing.T) {
 	assert.Equal(t, controltowerv1.EndpointKubernetesWorkloadKind_ENDPOINT_KUBERNETES_WORKLOAD_KIND_DEPLOYMENT, k8s.GetWorkloadKind())
 	assert.Equal(t, "checkout-75d84f5bdf-abc12", k8s.GetPodName())
 	assert.Equal(t, "pod-uid-123", k8s.GetPodUid())
-}
-
-func TestInvokingUserIgnoresSudoUserWhenNotElevated(t *testing.T) {
-	current, err := user.Current()
-	require.NoError(t, err)
-
-	orig := platform.IsPrivileged
-	t.Cleanup(func() { platform.IsPrivileged = orig })
-
-	// Non-root process: SUDO_USER must be ignored, else attribution is spoofable.
-	platform.IsPrivileged = func() bool { return false }
-	t.Setenv("SUDO_USER", "root")
-	got := invokingUser()
-	require.NotNil(t, got)
-	assert.Equal(t, current.Username, got.Username, "SUDO_USER must not override attribution when not elevated")
-
-	// Privileged: SUDO_USER is trusted and used.
-	platform.IsPrivileged = func() bool { return true }
-	t.Setenv("SUDO_USER", current.Username)
-	got = invokingUser()
-	require.NotNil(t, got)
-	assert.Equal(t, current.Username, got.Username)
-}
-
-func TestInvokingUserKeepsSudoAttributionWithoutPasswdEntry(t *testing.T) {
-	orig := platform.IsPrivileged
-	t.Cleanup(func() { platform.IsPrivileged = orig })
-
-	platform.IsPrivileged = func() bool { return true }
-	t.Setenv("SUDO_USER", "no-such-user-xyz")
-	t.Setenv("SUDO_UID", "4242")
-
-	got := invokingUser()
-	require.NotNil(t, got)
-	assert.Equal(t, "no-such-user-xyz", got.Username)
-	assert.Equal(t, "4242", got.Uid)
-
-	// Without SUDO_UID, fall back to the effective uid: a non-root username
-	// with uid 0 correctly signals the command ran under sudo.
-	t.Setenv("SUDO_UID", "")
-	got = invokingUser()
-	require.NotNil(t, got)
-	assert.Equal(t, "no-such-user-xyz", got.Username)
-	assert.Equal(t, "0", got.Uid)
 }
 
 func TestCloudSinkCapturesExecCommand(t *testing.T) {
